delete exception is implemented

+++ src/main/java/com/maveric/accountservice/services/AccountServiceImpl.java
@@ -1,17 +1,26 @@
 package com.maveric.accountservice.services;
 
+import com.maveric.accountservice.exception.AccountNotFoundException;
 import com.maveric.accountservice.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import static com.maveric.accountservice.enums.Constants.ACCOUNT_DELETED_SUCCESS;
+import static com.maveric.accountservice.enums.Constants.ACCOUNT_NOT_FOUND_MESSAGE;
 
 @Service
 public class AccountServiceImpl implements AccountService {
     @Autowired
     AccountRepository accountRepository;
+
     @Override
-    public String deleteAccount(String customerId, String accountId) {
+    public String deleteAccount(String accountId) {
+        if(!accountRepository.findById(accountId).isPresent())
+        {
+            throw new AccountNotFoundException(ACCOUNT_NOT_FOUND_MESSAGE+accountId);
+        }
         accountRepository.deleteById(accountId);
-        return "Account deleted successfully.";
+        return ACCOUNT_DELETED_SUCCESS;
     }
 }
